requirements.txt
Django==1.6.5
South==0.8.4
dj-database-url==0.3.0
dj-static==0.0.5
django-toolbelt==0.0.1
factory-boy==2.4.0
gunicorn==19.0.0
lxml==3.3.5
psycopg2==2.5.3
pystache==0.5.4
requests==2.3.0
static==1.0.2
wsgiref==0.1.2
